futher restructuring and placing files into folders
