+++ src/node/kademlia.go
@@ -1,6 +1,6 @@
 package node
 
-import "D7024E/contact"
+import "D7024E/node/contact"
 
 func LookupContact(target *contact.Contact) {
 	// TODO
